+++ src-ui/app/components/developer-debug-modal/misc-testing/misc-testing.component.ts
@@ -2,6 +2,7 @@ import { Component } from '@angular/core';
 import { VRChatMicMuteAutomationService } from '../../../services/osc-automations/vrchat-mic-mute-automation.service';
 import { invoke } from '@tauri-apps/api';
 import { OVRInputEventAction, OVRInputEventActionSet } from '../../../models/ovr-input-event';
+import { NotificationService, NotificationSound } from '../../../services/notification.service';
 
 @Component({
   selector: 'app-misc-testing',
@@ -10,8 +11,12 @@ import { OVRInputEventAction, OVRInputEventActionSet } from '../../../models/ovr
 })
 export class MiscTestingComponent {
   bindings: any;
+  soundVolume = 100;
 
-  constructor(protected automation: VRChatMicMuteAutomationService) {}
+  constructor(
+    protected automation: VRChatMicMuteAutomationService,
+    private notificationService: NotificationService
+  ) {}
 
   async openBindingConfig() {
     await invoke('openvr_launch_binding_configuration', { showOnDesktop: true });
@@ -28,14 +33,18 @@ export class MiscTestingComponent {
         actionSetKey: OVRInputEventActionSet.Main,
         actionKey: OVRInputEventAction.MuteMicrophone,
       }),
-      [OVRInputEventAction.SleepCheckDecline]: await invoke('openvr_get_binding_origins', {
+      [OVRInputEventAction.IndicatePresence]: await invoke('openvr_get_binding_origins', {
         actionSetKey: OVRInputEventActionSet.Hidden,
-        actionKey: OVRInputEventAction.SleepCheckDecline,
+        actionKey: OVRInputEventAction.IndicatePresence,
       }),
       [OVRInputEventAction.OverlayInteract]: await invoke('openvr_get_binding_origins', {
         actionSetKey: OVRInputEventActionSet.Hidden,
         actionKey: OVRInputEventAction.OverlayInteract,
       }),
     };
+  }
+
+  async playSound(sound: NotificationSound) {
+    await this.notificationService.playSound(sound, this.soundVolume / 100);
   }
 }