+++ src-ui/app/components/event-log-entry/event-log-entry.component.ts
@@ -22,6 +22,8 @@ import { EventLogWindowsPowerPolicySetEntryParser } from './entry-parsers/window
 import { EventLogMsiAfterburnerProfileSetEntryParser } from './entry-parsers/msi-afterburner-profile-set';
 import { EventLogSimpleBrightnessChangedEntryParser } from './entry-parsers/simple-brightness-changed';
 import { EventLogChangedVRChatMicMuteStateEntryParser } from './entry-parsers/changed-vrchat-mic-mute-state';
+import { EventLogChangedSystemMicControllerButtonBehaviorEntryParser } from './entry-parsers/changed-system-mic-controller-button-behavior';
+import { EventLogChangedSystemMicMuteStateEntryParser } from './entry-parsers/changed-system-mic-mute-state';
 
 @Component({
   selector: 'app-event-log-entry',
@@ -52,6 +54,8 @@ export class EventLogEntryComponent implements OnInit, OnChanges {
     new EventLogWindowsPowerPolicySetEntryParser(),
     new EventLogMsiAfterburnerProfileSetEntryParser(),
     new EventLogChangedVRChatMicMuteStateEntryParser(),
+    new EventLogChangedSystemMicMuteStateEntryParser(),
+    new EventLogChangedSystemMicControllerButtonBehaviorEntryParser(),
   ];
 
   constructor(