+++ src-ui/app/models/event-log-entry.ts
@@ -3,6 +3,7 @@ import { LighthouseDevicePowerState } from './lighthouse-device';
 import { SleepModeStatusChangeReason } from './sleep-mode';
 import { UserStatus } from 'vrchat/dist';
 import { WindowsPowerPolicy } from './windows-power-policy';
+import { SystemMicMuteControllerBindingBehavior } from './automations';
 
 export type EventLog = {
   version: 3;
@@ -32,6 +33,8 @@ export type EventLogEntry =
   | EventLogShutdownSequenceCancelled
   | EventLogWindowsPowerPolicySet
   | EventLogChangedVRChatMicMuteState
+  | EventLogChangedSystemMicMuteState
+  | EventLogChangedSystemMicControllerButtonBehavior
   | EventLogMsiAfterburnerProfileSet;
 
 export type EventLogDraft = Omit<EventLogEntry, 'time' | 'id'>;
@@ -54,6 +57,8 @@ export type EventLogType =
   | 'shutdownSequenceCancelled'
   | 'windowsPowerPolicySet'
   | 'changedVRChatMicMuteState'
+  | 'changedSystemMicMuteState'
+  | 'changedSystemMicControllerButtonBehavior'
   | 'msiAfterburnerProfileSet';
 
 export interface EventLogBase {
@@ -175,5 +180,18 @@ export interface EventLogMsiAfterburnerProfileSet extends EventLogBase {
 export interface EventLogChangedVRChatMicMuteState extends EventLogBase {
   type: 'changedVRChatMicMuteState';
   muted: boolean;
+  reason: 'SLEEP_MODE_ENABLED' | 'SLEEP_MODE_DISABLED' | 'SLEEP_PREPARATION';
+}
+
+export interface EventLogChangedSystemMicMuteState extends EventLogBase {
+  type: 'changedSystemMicMuteState';
+  muted: boolean;
+  deviceName: string;
+  reason: 'SLEEP_MODE_ENABLED' | 'SLEEP_MODE_DISABLED' | 'SLEEP_PREPARATION';
+}
+
+export interface EventLogChangedSystemMicControllerButtonBehavior extends EventLogBase {
+  type: 'changedSystemMicControllerButtonBehavior';
+  behavior: 'TOGGLE' | 'PUSH_TO_TALK';
   reason: 'SLEEP_MODE_ENABLED' | 'SLEEP_MODE_DISABLED' | 'SLEEP_PREPARATION';
 }