+++ src-ui/app/services/power-automations/turn-off-devices-on-battery-level-automation.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { cloneDeep } from 'lodash';
 import { AutomationConfigService } from '../automation-config.service';
 import { OpenVRService } from '../openvr.service';
-import { map } from 'rxjs';
+import { combineLatest, map } from 'rxjs';
 import {
   AUTOMATION_CONFIGS_DEFAULT,
   TurnOffDevicesOnBatteryLevelAutomationConfig,
@@ -12,6 +12,7 @@ import { OVRDevice, OVRDeviceClass } from '../../models/ovr-device';
 import { LighthouseConsoleService } from '../lighthouse-console.service';
 import { error } from 'tauri-plugin-log-api';
 import { EventLogTurnedOffOpenVRDevices } from '../../models/event-log-entry';
+import { SleepService } from '../sleep.service';
 
 @Injectable({
   providedIn: 'root',
@@ -31,7 +32,8 @@ export class TurnOffDevicesOnBatteryLevelAutomationService {
     private automationConfig: AutomationConfigService,
     private openvr: OpenVRService,
     private lighthouse: LighthouseConsoleService,
-    private eventLog: EventLogService
+    private eventLog: EventLogService,
+    private sleep: SleepService
   ) {}
 
   async init() {
@@ -39,13 +41,13 @@ export class TurnOffDevicesOnBatteryLevelAutomationService {
       .pipe(map((configs) => configs.TURN_OFF_DEVICES_ON_BATTERY_LEVEL))
       .subscribe((config) => (this.config = config));
 
-    this.openvr.devices.subscribe((devices) => {
+    combineLatest([this.openvr.devices, this.sleep.mode]).subscribe(([devices, sleepMode]) => {
       for (const device of devices) {
         const previousLevel = this.batteryLevelCache[device.index]?.level || null;
         const currentLevel = device.battery;
         this.batteryLevelCache[device.index] = { level: currentLevel, class: device.class };
         if (!currentLevel || currentLevel === previousLevel || device.isCharging) return;
-        this.processBatteryChange(device, previousLevel, currentLevel);
+        this.processBatteryChange(device, previousLevel, currentLevel, sleepMode);
       }
     });
   }
@@ -53,7 +55,8 @@ export class TurnOffDevicesOnBatteryLevelAutomationService {
   private async processBatteryChange(
     device: OVRDevice,
     previousLevel: number | null,
-    currentLevel: number
+    currentLevel: number,
+    sleepMode: boolean
   ) {
     if (previousLevel === null || previousLevel <= currentLevel) return;
     if (!['Controller', 'Tracker'].includes(device.class)) return;
@@ -61,11 +64,13 @@ export class TurnOffDevicesOnBatteryLevelAutomationService {
     switch (device.class) {
       case 'Controller':
         if (!this.config.turnOffControllers) return;
+        if (this.config.turnOffControllersOnlyDuringSleepMode && !sleepMode) return;
         if (currentLevel > this.config.turnOffControllersAtLevel) return;
         threshold = this.config.turnOffControllersAtLevel;
         break;
       case 'GenericTracker':
         if (!this.config.turnOffTrackers) return;
+        if (this.config.turnOffTrackersOnlyDuringSleepMode && !sleepMode) return;
         if (currentLevel > this.config.turnOffTrackersAtLevel) return;
         threshold = this.config.turnOffTrackersAtLevel;
         break;